requirements.txt
pyTelegramBotAPI
requests
beautifulsoup4
python-dotenv